+++ R/FilteringMethods.R
@@ -13,14 +13,15 @@
 #' @param pVal Used only if method = "poisson". Default value is 0.01.
 #' @param tpmLow Used only if method = "TPM". Default value is 0.1.
 #' @return A modified TSSr object with updated \code{TSSprocessedMatrix}
-#'   slot after filtering.
+#'   slot after filtering. The input object is not modified; assign the
+#'   returned object to retain the changes.
 #'
 #'
 #' @export
 #'
 #' @examples
 #' data(exampleTSSr)
-#' filterTSS(exampleTSSr, method = "TPM", tpmLow = 0.1)
+#' exampleTSSr <- filterTSS(exampleTSSr, method = "TPM", tpmLow = 0.1)
 setGeneric("filterTSS", function(
   object, method = "poisson", normalization = TRUE,
   pVal = 0.01, tpmLow = 0.1
@@ -32,7 +33,6 @@ setGeneric("filterTSS", function(
 setMethod("filterTSS", signature(object = "TSSr"), function(object, method, normalization, pVal, tpmLow) {
     ## initialize values
     sampleLabelsMerged <- object@sampleLabelsMerged
-    objName <- deparse(substitute(object))
     library.size <- object@librarySizes
 
     tss.dt <- object@TSSprocessedMatrix
@@ -73,7 +73,6 @@ setMethod("filterTSS", signature(object = "TSSr"), function(object, method, norm
         re <- re[rowSums(re[, 4:ncol(re)]) > 0, ]
         setorder(re, "strand", "chr", "pos")
         object@TSSprocessedMatrix <- re
-        assign(objName, object, envir = parent.frame())
     } else if (method == "TPM") {
         message("\nFiltering data with ", method, " method...")
         if (!all(is.normalized)) {
@@ -95,8 +94,8 @@ setMethod("filterTSS", signature(object = "TSSr"), function(object, method, norm
         re <- re[rowSums(re[, 4:ncol(re)]) > 0, ]
         setorder(re, "strand", "chr", "pos")
         object@TSSprocessedMatrix <- re
-        assign(objName, object, envir = parent.frame())
     } else {
         message("\tNo filtering method is defined...")
     }
+    return(object)
 })