+++ R/MergingMethods.R
@@ -7,17 +7,17 @@
 #' @param object A TSSr object
 #' @param mergeIndex Integer vector specifying which samples to be merged
 #' @return A modified TSSr object with updated \code{TSSprocessedMatrix}
-#'   and \code{librarySizes} slots after merging samples.
+#'   and \code{librarySizes} slots after merging samples. The input object is
+#'   not modified; assign the returned object to retain the changes.
 #' @export
 #'
 #' @examples
 #' data(exampleTSSr)
-#' mergeSamples(exampleTSSr, mergeIndex = c(1, 1, 2, 2))
+#' exampleTSSr <- mergeSamples(exampleTSSr, mergeIndex = c(1, 1, 2, 2))
 setGeneric("mergeSamples", function(object, mergeIndex = NULL) standardGeneric("mergeSamples"))
 #' @rdname mergeSamples
 #' @export
 setMethod("mergeSamples", signature(object = "TSSr"), function(object, mergeIndex) {
-    objName <- deparse(substitute(object))
     if (is.null(mergeIndex)) {
         mergeIndex <- as.integer(object@mergeIndex)
     } else {
@@ -49,5 +49,5 @@ setMethod("mergeSamples", signature(object = "TSSr"), function(object, mergeInde
     # object@mergeIndex <- mergeIndex
     object@TSSprocessedMatrix <- re
     object@librarySizes <- colSums(re[, 4:ncol(re), drop = FALSE], na.rm = TRUE)
-    assign(objName, object, envir = parent.frame())
+    return(object)
 })